Push to main
"This was my last major pass on the story and game actions. Like previous ones I have added few more words that could be used for certain actions. I have made massive changes to the psych room actions and descriptions to make it more clear to the user as to what is going on. I have made a few changes to the texts and actions regarding the database computer to hopefully make it more clear in terms of using it. I added one more ending regarding the gate and guard post situation in the yard scene. I made couple changes to attaining and losing items to hopefully make things a bit more clear. @everyone We still need lambdas and more comments on code hopefully you guys can cover this (some chunks of code have little to no comments and we need to mark our code with comments to show that certain requirements where met). I need to focus on other school work now so I won't be making any changes unless they are necessary. @everyone Also we should all meet on Tuesday to talk about our final project and wrap things up. Otherwise please check that everything is good and try to finish these last requirements."